Conway's Game of Life in Haskell
The famous Game of Life is an cellular automaton simulation created by
mathematician John Conway.
It involves a series of states in which each pixel on a grid is "alive"
(filled in) or "dead" (empty) based on the previous state. Using a small
set of rules, the user gives an initial state which often creates
surprisingly intricate and amazing geometrical patterns. Though the
"game" only involves initial interaction by the user, it's definitely
still a lot of fun!

We used the functional language Haskell to recreate the game. We added a
few features which show information of interest to the user, such as
tracing which pixels have been filled and showing the number of alive
cells at any given time. A list of all the functions can be found in the
readme.